New capability for running il2cpp and linker with bee
* Using the bee build graph to detect situations where there are zero
code
changes and il2cpp and linker do not need to be run.
* Adding build performance tests.
* Turning on response files for unity linker options.  Fixed an issue
with response files having multiple arguments on the same file in unity
options.
* Leaving mscrolib references when clearing all references using
predictable linker output when making an empty assembly for one that
would have been removed.  Windows Runtime projection code expects to be
able to find corlib and will throw exceptions when processing assemblies
that do not reference it.
* Handling no-stripping case for Bee and non-Bee build performance
tests.